backend/requirements.txt
flask
flask-cors
redis
kafka-python
mysql-connector-python
werkzeug

# OpenTelemetry core
opentelemetry-api
opentelemetry-sdk

# OTLP Exporters (HTTP 사용)
opentelemetry-exporter-otlp-proto-http

# 자동 계측
opentelemetry-instrumentation
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-redis
opentelemetry-instrumentation-pymysql
opentelemetry-instrumentation-logging 